+++ prestaspot/blocks/product-list/index.asset.php
@@ -8,5 +8,5 @@ return array(
         'wp-server-side-render',
         'wp-i18n',
     ),
-    'version' => '0.8.0',
+    'version' => '0.9.0',
 );